+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
+## Unreleased
+- Album preview in invite page
 
 ## 2024-02-19
 - Move wedding link to header